Add files via upload
Source code.  It's all one big file, in python3.  Just download it somewhere in your path, read the description in the comments near the top, and see if it does what you need.   It is in python3, and has rather vanilla-style dependencies -- astropy.io.fits, numpy, matplotlib, and a few standard python modules such as datetiime and os.