Add files via upload
This is a short config file that specifies various numbers for running the code, to be placed in the same directory as your data (or pointed to with the -p command line argument).  This particular file is optimized for OSMOS data taken with the object positioned consistently on the slit; you'll need to adjust various numbers for your own setup.  The format is obvious.  Read the description at the top of the source code, which describes what most of these do.